Extract fed resolution tests
